logicaffeine_compile::parser

Trait LogicVerbParsing

pub trait LogicVerbParsing<'a, 'ctx, 'int> {
    // Required methods
    fn parse_predicate_with_subject(
        &mut self,
        subject_symbol: Symbol,
    ) -> Result<&'a LogicExpr<'a>, ParseError>;
    fn parse_predicate_with_subject_as_var(
        &mut self,
        subject_symbol: Symbol,
    ) -> Result<&'a LogicExpr<'a>, ParseError>;
    fn try_parse_plural_subject(
        &mut self,
        first_subject: &NounPhrase<'a>,
    ) -> Result<Option<&'a LogicExpr<'a>>, ParseError>;
    fn parse_control_structure(
        &mut self,
        subject: &NounPhrase<'a>,
        verb: Symbol,
        verb_time: Time,
    ) -> Result<&'a LogicExpr<'a>, ParseError>;
    fn is_control_verb(&self, verb: Symbol) -> bool;
    fn build_group_predicate(
        &mut self,
        subjects: &[Symbol],
        verb: Symbol,
        verb_time: Time,
    ) -> &'a LogicExpr<'a>;
    fn build_group_transitive(
        &mut self,
        subjects: &[Symbol],
        objects: &[Symbol],
        verb: Symbol,
        verb_time: Time,
    ) -> &'a LogicExpr<'a>;
}

Required Methods

fn parse_predicate_with_subject( &mut self, subject_symbol: Symbol, ) -> Result<&'a LogicExpr<'a>, ParseError>

fn parse_predicate_with_subject_as_var( &mut self, subject_symbol: Symbol, ) -> Result<&'a LogicExpr<'a>, ParseError>

fn try_parse_plural_subject( &mut self, first_subject: &NounPhrase<'a>, ) -> Result<Option<&'a LogicExpr<'a>>, ParseError>

Try to parse a plural subject construction like “John and Mary verb”. Returns Ok(Some(expr)) if successful, Ok(None) if not a plural subject (backtrack), or Err if there’s a semantic error (e.g., “respectively” with mismatched lengths).

fn parse_control_structure( &mut self, subject: &NounPhrase<'a>, verb: Symbol, verb_time: Time, ) -> Result<&'a LogicExpr<'a>, ParseError>

fn is_control_verb(&self, verb: Symbol) -> bool

fn build_group_predicate( &mut self, subjects: &[Symbol], verb: Symbol, verb_time: Time, ) -> &'a LogicExpr<'a>

Build a group predicate for intransitive verbs with multiple subjects

fn build_group_transitive( &mut self, subjects: &[Symbol], objects: &[Symbol], verb: Symbol, verb_time: Time, ) -> &'a LogicExpr<'a>

Build a transitive predicate with group subject and group object